requirements.txt
flask
flask-cors
gunicorn
pymongo
python-dotenv
pycoingecko
numpy